Add XcodeProj dependency and fix ProjectSettings and XcodeProject
There were duplicated model ProjectSettings and XcodeProject -- do
same thing. Replace it by XcodeProjectList
