+++ Sources/DepoCore/PackageManagers/PodManager.swift
@@ -159,7 +159,7 @@ public final class PodManager: ProgressObservable {
     }
 
     private func allSchemes() throws -> [(Pod, BuildSettings)] {
-        let project = try XcodeProject(shell: shell)
+        let project = try XcodeProjectList(shell: shell)
         return try project.targets.compactMap { targetName in
             guard !targetName.starts(with: podsInternalTargetsPrefix) else {
                 return nil
